# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Employee Directory & Branch Management API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the Employee Directory & Branch Management Application."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Employee": {
        "type": "object",
        "required": [
          "id",
          "name",
          "position",
          "department",
          "email",
          "phone",
          "branchId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the employee",
            "example": "HC3mXN6Dxyg4eaSjhBxY"
          },
          "name": {
            "type": "string",
            "description": "Full name of the employee",
            "example": "John Black"
          },
          "position": {
            "type": "string",
            "description": "Job position or title of the employee",
            "example": "Branch Manager"
          },
          "department": {
            "type": "string",
            "description": "Department where the employee works",
            "example": "Management"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Employee's email address",
            "example": "john.black@pixell-river.com"
          },
          "phone": {
            "type": "string",
            "description": "Employee's phone number",
            "example": "204-555-04667"
          },
          "branchId": {
            "type": "string",
            "description": "Identifier for the branch where employee is located",
            "example": "6"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/employee": {
      "get": {
        "summary": "Retrieves a list of employees",
        "tags": [
          "Employee"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of Employees",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new employee",
        "tags": [
          "Employee"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "position",
                  "department",
                  "email",
                  "phone",
                  "branchId"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Black"
                  },
                  "position": {
                    "type": "string",
                    "example": "Branch Manager"
                  },
                  "department": {
                    "type": "string",
                    "example": "Management"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john.black@pixell-river.com"
                  },
                  "phone": {
                    "type": "string",
                    "example": "204-555-04667"
                  },
                  "branchId": {
                    "type": "string",
                    "example": "6"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Name is required"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/employee/{id}": {
      "get": {
        "summary": "Retrieves a single employee by ID",
        "tags": [
          "Employee"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The employee ID",
            "example": "HC3mXN6Dxyg4eaSjhBxY"
          }
        ],
        "responses": {
          "201": {
            "description": "Employee details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "400": {
            "description": "Missing employee ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Employee ID is required"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Employee not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an existing employee",
        "tags": [
          "Employee"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The employee ID",
            "example": "HC3mXN6Dxyg4eaSjhBxY"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Gray"
                  },
                  "position": {
                    "type": "string",
                    "example": "Branch Manager"
                  },
                  "department": {
                    "type": "string",
                    "example": "Management"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john.black@pixell-river.com"
                  },
                  "phone": {
                    "type": "string",
                    "example": "204-555-04667"
                  },
                  "branchId": {
                    "type": "string",
                    "example": "6"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "400": {
            "description": "Invalid employee ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Employee ID is required"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Employee not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an employee",
        "tags": [
          "Employee"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The employee ID",
            "example": "HC3mXN6Dxyg4eaSjhBxY"
          }
        ],
        "responses": {
          "200": {
            "description": "Employee deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Employee deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid employee ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Employee ID is required"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Employee not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/employee/branch/{branchId}": {
      "get": {
        "summary": "Retrieves all employees for a specific branch",
        "tags": [
          "Employee"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "branchId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The branch ID",
            "example": "6"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of employees for the specified branch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                },
                "example": [
                  {
                    "id": "HC3mXN6Dxyg4eaSjhBxY",
                    "name": "John Black",
                    "position": "Branch Manager",
                    "department": "Management",
                    "email": "john.black@pixell-river.com",
                    "phone": "204-555-04667",
                    "branchId": "6"
                  },
                  {
                    "id": "ABC123XYZ456",
                    "name": "Sarah Johnson",
                    "position": "Senior Developer",
                    "department": "Engineering",
                    "email": "sarah.johnson@pixell-river.com",
                    "phone": "204-555-01234",
                    "branchId": "6"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Invalid branch ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Branch ID is required"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Branch not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Branch not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/employee/department/:departmentName": {
      "get": {
        "summary": "Retrieves all employees for a specific department",
        "tags": [
          "Employee"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "department",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The department name",
            "example": "Management"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of employees for the specified Department",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Department Name",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Department Name is required"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Department not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Department not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}